chomp get api and other minor fixes
new chomp get api, added chomp json processing logic, added pod-site
association logic, minor fixes to pod, region params

+++ AECPortalMgmt/src/main/java/org/akraino/portal/daoimpl/RegionDAOImpl.java
@@ -64,7 +64,7 @@ public class RegionDAOImpl implements RegionDAO {
 	}
 
 	@Override
-	public Region getRegion(Long regionId) {
+	public Region getRegion(Integer regionId) {
 
 		EntityManager em = getSession().getEntityManagerFactory().createEntityManager();
